+++ src/xxhash.c
@@ -1,0 +1,2 @@
+#define XXH_IMPLEMENTATION
+#include "xxhash.h"